Next generation ISR handling

+++ libgloss/mips/uhi_read.c
@@ -3,7 +3,7 @@
 */
 
 /*
- * Copyright (c) 2014, Imagination Technologies Ltd.
+ * Copyright (c) 2015, Imagination Technologies Ltd.
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -32,34 +32,27 @@
 */
 
 /*
- * @Synopsis     int32_t read (int32_t fd, void *buffer, int32_t len);
+ * @Synopsis	 read
  *
- *               Parameters:
- *                 $4 - File handle
- *                 $5 - Destination buffer
- *                 $6 - Max number of bytes to read
+ *		 Arguments to syscall:
+ *		   $25 - Operation code for read
+ *		   $4 - File handle
+ *		   $5 - Destination buffer
+ *		   $6 - Max number of bytes to read
  *
- *               Return:
- *                 $2 - Number of bytes read
- *
- *               Arguments to syscall:
- *                 $25 - Operation code for read
- *                 $4 - File handle
- *                 $5 - Destination buffer
- *                 $6 - Max number of bytes to read
- *
- *               Return from syscall:
- *                 $2 - Number of bytes read
- *                 $3 - errno
+ *		 Return from syscall:
+ *		   $2 - Number of bytes read
+ *		   $3 - errno
  *
  * @Description  File read
 */
 
 #include <stdint.h>
 #include <errno.h>
-#include "uhi_syscalls.h"
+#include <mips/uhi_syscalls.h>
 
-int32_t read (int32_t fd, void *buffer, int32_t len)
+int
+read (int fd, void *buffer, size_t len)
 {
   register int32_t arg1 asm ("$4") = fd;
   register void *arg2 asm ("$5") = buffer;
@@ -68,19 +61,19 @@ int32_t read (int32_t fd, void *buffer, int32_t len)
   register int32_t ret asm ("$2") = __MIPS_UHI_SYSCALL_NUM;
   register int32_t new_errno asm ("$3") = 0;
 
-  __asm__ __volatile__(" # %0,%1 = read(%2, %3, %4) op=%5\n"
-                       SYSCALL (__MIPS_UHI_SYSCALL_NUM)
-                       : "+r" (ret), "=r" (new_errno), "+r" (arg1), "+r" (arg2)
-		       : "r" (arg3), "r" (op));
+  __asm__ __volatile__ (" # %0,%1 = read(%2, %3, %4) op=%5\n"
+			SYSCALL (__MIPS_UHI_SYSCALL_NUM)
+			: "+r" (ret), "=r" (new_errno), "+r" (arg1),
+			  "+r" (arg2)
+			: "r" (arg3), "r" (op));
 
   if (ret == -1)
     {
       /* Do a dance to set errno, errno is a function call that can
-         clobber $3.  */
+	 clobber $3.  */
       volatile uint32_t errno_tmp = new_errno;
       errno = errno_tmp;
     }
 
   return ret;
 }
-
